new updatee

+++ src/main/java/com/uitm/myattend/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.uitm.myattend.controller;
 
+import com.uitm.myattend.model.CommonModel;
 import com.uitm.myattend.model.UserModel;
 import com.uitm.myattend.service.AuthService;
 import com.uitm.myattend.service.StudentService;
@@ -8,12 +9,14 @@ import com.uitm.myattend.utility.FieldUtility;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.TransactionCallbackWithoutResult;
 import org.springframework.transaction.support.TransactionTemplate;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.net.http.HttpResponse;
@@ -40,6 +43,11 @@ public class UserController {
             response.sendRedirect(request.getContextPath() + "/login");
             return null;
         }
+
+        if(!authService.authorize(session, FieldUtility.ADMIN_ROLE)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         List<UserModel> userList = userService.retrieveAll(session);
         request.setAttribute("users", userList);
         request.setAttribute("totalUser", userList.size());
@@ -57,6 +65,10 @@ public class UserController {
             if(!authService.authenticate(session)) {
                 response.sendRedirect(request.getContextPath() + "/login");
                 return;
+            }
+
+            if(!authService.authorize(session, FieldUtility.ADMIN_ROLE)) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
             }
 
             FieldUtility.requiredValidator(body, userRequiredFields());
@@ -94,6 +106,13 @@ public class UserController {
             if(!authService.authenticate(session)) {
                 respMap.put("respCode", "00002");
                 respMap.put("respStatus", "error");
+                respMap.put("respMessage", "Unauthenticated request");
+                return respMap;
+            }
+
+            if(!authService.authorize(session, FieldUtility.ADMIN_ROLE)) {
+                respMap.put("respCode", "00099");
+                respMap.put("respStatus", "error");
                 respMap.put("respMessage", "Unauthorized request");
                 return respMap;
             }
@@ -129,6 +148,10 @@ public class UserController {
                 return;
             }
 
+            if(!authService.authorize(session, FieldUtility.ADMIN_ROLE)) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+
             FieldUtility.requiredValidator(body, userRequiredFields());
 
             if(!userService.update(body, file)) {
@@ -144,12 +167,81 @@ public class UserController {
     }
 
     @PostMapping("/delete")
-    public void delete(@RequestParam Map<String, Object> body, HttpServletResponse response, HttpSession session) throws IOException {
+    public void delete(@RequestParam Map<String, Object> body, HttpServletResponse response, HttpServletRequest request, HttpSession session) throws IOException {
         try {
+            if(!authService.authenticate(session)) {
+                response.sendRedirect(request.getContextPath() + "/login");
+                return;
+            }
+
+            if(!authService.authorize(session, FieldUtility.ADMIN_ROLE)) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+
             if(!userService.delete(body)) {
                 throw new Exception("Failed to delete user data");
             }else {
                 session.setAttribute("success", "User data successfully deleted");
+            }
+        }catch (Exception e) {
+            session.setAttribute("error", e.getMessage());
+            e.printStackTrace();
+        }
+        response.sendRedirect("/user");
+    }
+
+    @GetMapping("/profile")
+    @ResponseBody
+    public Map<String, Object> profile(HttpServletResponse response, HttpServletRequest request, HttpSession session) {
+        Map<String, Object> respMap = new HashMap<>();
+        try {
+            if(!authService.authenticate(session)) {
+                respMap.put("respCode", "00002");
+                respMap.put("respStatus", "error");
+                respMap.put("respMessage", "Unauthenticated request");
+                return respMap;
+            }
+
+
+            CommonModel common = (CommonModel) session.getAttribute("common");
+            UserModel user = userService.retrieveUserById(common.getUser().getId());
+
+            if(user == null) {
+                respMap.put("respCode", "00001");
+                respMap.put("respStatus", "error");
+                respMap.put("respMessage", "User does not found!");
+            }else{
+                respMap.put("respCode", "00000");
+                respMap.put("respStatus", "success");
+                respMap.put("respMessage", "successfully retrieved");
+            }
+            respMap.put("data", user);
+        }catch (Exception e) {
+            e.printStackTrace();
+            //session.setAttribute("message", "Internal server error. Please contact admin for futher assistance");
+            respMap.put("respCode", "000198");
+            respMap.put("respStatus", "error");
+            respMap.put("respMessage", "Internal server error. Please contact admin for futher assistance");
+        }
+        return respMap;
+    }
+
+    @PostMapping("/profile")
+    public void profileUpdate(@RequestParam Map<String, Object> body, @RequestParam("dpImage") MultipartFile file, HttpServletResponse response, HttpServletRequest request ,HttpSession session) throws IOException {
+        try {
+            if(!authService.authenticate(session)) {
+                response.sendRedirect(request.getContextPath() + "/login");
+                return;
+            }
+
+
+            CommonModel common = (CommonModel) session.getAttribute("common");
+            FieldUtility.requiredValidator(body, userRequiredFields());
+            body.put("uid", common.getUser().getId());
+            if(!userService.update(body, file)) {
+                throw new Exception("Failed to update profile");
+            }else {
+                session.setAttribute("success", "Profile updated successfully");
             }
         }catch (Exception e) {
             session.setAttribute("error", e.getMessage());
